additional update 2021-03-20
Combined Items: further attempts made to use CompoundTechs to unlock blueprints when multiple others are unlocked. Progress has been made; the CompoundTechs themselves work. However, when attempting to add them in PostPatch, or in a coroutine triggered in PostPatch, the coroutine is prematurely stopped before KnownTech is initialised, and thus the compoundTechs are never added without intervention with Runtime Editor